Closes #4: show status when only one app is active
Previously, pressing alt + tab when only one app was open in a workspace would behave exactly the same as when there were no apps open (i.e., it would do nothing).

With this change (which basically just removes the conditional that was creating the above-mentioned behaviour), it now shows the app that is open by itself in the switcher. This is a reinforcing behaviour and is also consistent with how the switcher behaves otherwise.